top level env created and minus AppC also working

+++ Value.cr
@@ -4,7 +4,7 @@ abstract class Val
 end
 # NumV
 class NumV < Val
-    getter n : Int32
+    getter n : Float64
     def initialize(@n)
     end
